# kanban/openapi.yaml
openapi: 3.1.0
info:
  title: TaskFlow Kanban API
  version: 2.1.0
  description: |
    Backend API for the TaskFlow Kanban MAX mini-application.
    Unless stated otherwise, API operations require valid MAX WebApp
    initialization data supplied in the X-Max-InitData header.
servers:
  - url: https://kirillgvoz.ru
    description: Production deployment
  - url: http://localhost:9300
    description: Local Docker deployment
security:
  - MaxInitData: []
components:
  securitySchemes:
    MaxInitData:
      type: apiKey
      in: header
      name: X-Max-InitData
      description: Signed MAX WebApp initialization data.
    MaxWebhookSecret:
      type: apiKey
      in: header
      name: X-Max-Bot-Api-Secret
      description: Shared webhook secret configured on both MAX and TaskFlow.
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    Ok:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
...
    OrgMember:
      type: object
      required: [org_id, user_id, role, joined_at]
      properties:
        org_id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        username:
          type: string
        display_name:
          type: string
        role:
          type: string
          enum: [owner, admin, member]
        joined_at:
          type: string
          format: date-time
...
    Board:
      type: object
      required: [id, org_id, name, is_archived, created_by, created_at]
      properties:
        id:
          type: integer
          format: int64
        org_id:
          type: integer
          format: int64
        name:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
        is_archived:
          type: boolean
        created_by:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
    BoardDetail:
      allOf:
        - $ref: '#/components/schemas/Board'
        - type: object
          required: [columns, members]
          properties:
            columns:
              type: array
              items:
                $ref: '#/components/schemas/Column'
            members:
              type: array
              items:
                $ref: '#/components/schemas/OrgMember'
    BoardColumnInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 80
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
    BoardCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
          maxLength: 2000
        columns:
          type: array
          maxItems: 30
          items:
            $ref: '#/components/schemas/BoardColumnInput'
    BoardUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
          maxLength: 2000
        is_archived:
          type: boolean
    Column:
      type: object
      required: [id, board_id, name, position, color]
      properties:
        id:
          type: integer
          format: int64
        board_id:
          type: integer
          format: int64
        name:
          type: string
        position:
          type: integer
          minimum: 0
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
...
    Assignee:
      type: object
      required: [task_id, user_id]
      properties:
        task_id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        username:
          type: string
        display_name:
          type: string
    Task:
      type: object
      required: [id, board_id, column_id, title, position, priority, created_by, created_at, updated_at]
      properties:
        id:
          type: integer
          format: int64
        board_id:
          type: integer
          format: int64
        column_id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        position:
          type: integer
          minimum: 0
        priority:
          type: string
          enum: [low, medium, high, urgent]
        deadline:
          type: [string, 'null']
          pattern: '^[0-9]{4}-[0-9]{2}-[0-9]{2}$'
        created_by:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        assignees:
          type: array
          items:
            $ref: '#/components/schemas/Assignee'
    ChecklistItem:
      type: object
      required: [id, checklist_id, text, is_done, position]
      properties:
        id:
          type: integer
          format: int64
        checklist_id:
          type: integer
          format: int64
        text:
          type: string
        is_done:
          type: boolean
        position:
          type: integer
          minimum: 0
    Checklist:
      type: object
      required: [id, task_id, title, position, items]
      properties:
        id:
          type: integer
          format: int64
        task_id:
          type: integer
          format: int64
        title:
          type: string
        position:
          type: integer
          minimum: 0
        items:
          type: array
          items:
            $ref: '#/components/schemas/ChecklistItem'
    Comment:
      type: object
      required: [id, task_id, user_id, text, created_at]
      properties:
        id:
          type: integer
          format: int64
        task_id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        username:
          type: string
        display_name:
          type: string
        text:
          type: string
        created_at:
          type: string
          format: date-time
    TaskDetail:
      allOf:
        - $ref: '#/components/schemas/Task'
        - type: object
          required: [checklists, comments]
          properties:
            checklists:
              type: array
              items:
                $ref: '#/components/schemas/Checklist'
            comments:
              type: array
              items:
                $ref: '#/components/schemas/Comment'
    TaskCreate:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 300
        description:
          type: string
          maxLength: 10000
        column_id:
          type: integer
          format: int64
          minimum: 1
        priority:
          type: string
          enum: [low, medium, high, urgent]
          default: medium
        deadline:
          type: [string, 'null']
          pattern: '^[0-9]{4}-[0-9]{2}-[0-9]{2}$'
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 300
        description:
          type: string
          maxLength: 10000
        priority:
          type: string
          enum: [low, medium, high, urgent]
        deadline:
          type: [string, 'null']
          pattern: '^[0-9]{4}-[0-9]{2}-[0-9]{2}$'
    TaskMove:
      type: object
      required: [column_id, position]
      properties:
        column_id:
          type: integer
          format: int64
          minimum: 1
        position:
          type: integer
          minimum: 0
    TaskAssign:
      type: object
      required: [user_id]
      properties:
        user_id:
          type: integer
          format: int64
          minimum: 1
...
paths:
...
  /max-kanban/api/orgs/{id}/boards:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
    get:
      operationId: listBoards
      summary: List organization boards.
      parameters:
        - name: include_archived
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Boards.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Board'
        '401':
          description: Missing or invalid authentication.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Organization not found or not accessible.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      operationId: createBoard
      summary: Create a board with optional custom statuses.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BoardCreate'
      responses:
        '201':
          description: Board created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Board'
        '400':
          description: Invalid board or status data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing or invalid authentication.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Admin or owner role required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /max-kanban/api/boards/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
    get:
      operationId: getBoard
      summary: Get a board with statuses and members.
      responses:
        '200':
          description: Board detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardDetail'
        '401':
          description: Missing or invalid authentication.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Board not found or not accessible.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      operationId: updateBoard
      summary: Update or archive a board.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BoardUpdate'
      responses:
        '200':
          description: Board updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
        '400':
          description: Invalid board data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing or invalid authentication.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Admin or owner role required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Board not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      operationId: archiveBoard
      summary: Archive a board without deleting its history.
      responses:
        '200':
          description: Board archived.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
        '401':
          description: Missing or invalid authentication.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Admin or owner role required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Board not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /max-kanban/api/boards/{id}/tasks:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
    get:
      operationId: listTasks
      summary: List board tasks.
      responses:
        '200':
          description: Tasks.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          description: Missing or invalid authentication.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Board not found or not accessible.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      operationId: createTask
      summary: Create a task on a board.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid task or status data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing or invalid authentication.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Board not found or not accessible.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Board has no status for new tasks.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /max-kanban/api/tasks/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
    get:
      operationId: getTask
      summary: Get a task with checklists and comments.
      responses:
        '200':
          description: Task detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDetail'
        '401':
          description: Missing or invalid authentication.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Task not found or not accessible.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      operationId: updateTask
      summary: Update task fields.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid task data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing or invalid authentication.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Task not found or not accessible.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      operationId: deleteTask
      summary: Delete a task.
      responses:
        '200':
          description: Task deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
        '401':
          description: Missing or invalid authentication.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Task not found or not accessible.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /max-kanban/api/tasks/{id}/move:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
    put:
      operationId: moveTask
      summary: Move a task to another status and position.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskMove'
      responses:
        '200':
          description: Task moved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid status or position.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing or invalid authentication.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Task not found or not accessible.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /max-kanban/api/tasks/{id}/assign:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
    post:
      operationId: assignTask
      summary: Assign an organization member to a task.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskAssign'
      responses:
        '200':
          description: Task assigned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assignee'
        '400':
          description: Invalid assignee.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing or invalid authentication.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Task not found or not accessible.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...